resource: refactor PCM library into its own module
