+++ source/vstd/tokens/map.rs
@@ -76,8 +76,10 @@
 use super::super::map::*;
 use super::super::map_lib::*;
 use super::super::modes::*;
-use super::super::pcm::*;
 use super::super::prelude::*;
+use super::super::resource::Loc;
+use super::super::resource::Resource;
+use super::super::resource::pcm::PCM;
 use super::super::set_lib::*;
 
 verus! {
