+++ source/vstd/vstd.rs
@@ -56,13 +56,12 @@ pub mod math;
 pub mod modes;
 pub mod multiset;
 pub mod multiset_lib;
-pub mod pcm;
-pub mod pcm_lib;
 pub mod pervasive;
 pub mod predicate;
 pub mod proph;
 pub mod raw_ptr;
 pub mod relations;
+pub mod resource;
 pub mod rwlock;
 pub mod seq;
 pub mod seq_lib;
